#9: Refactor code
- divide it into some compact files: done
